+++ firearms/src/FirearmsMod.cs
@@ -15,10 +15,9 @@ namespace MaltiezFirearms
         public override void Start(ICoreAPI api)
         {
             base.Start(api);
-            api.RegisterItemClass("MaltiezFirearm", typeof(ItemFirearm));
             api.RegisterItemClass("MaltiezPowderFlask", typeof(ItemPowderFlask)); 
             api.RegisterItemClass("MaltiezMatch", typeof(ItemMatch));
-            api.RegisterItemClass("MaltiezMusket", typeof(ItemMusket));
+            api.RegisterItemClass("FirearmsNoMelee", typeof(NoMelee));
         }
     }
 }
